LIBFCREPO-1359. Rearranged test directories for app testing.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1359
